chore(be): organize controllers
