--- backend/apps/client/src/contest/dto/respond-publicizing-request.dto.ts
@@ -1,7 +1,0 @@
-import { IsBoolean, IsNotEmpty } from 'class-validator'
-
-export class RespondContestPublicizingRequestDto {
-  @IsBoolean()
-  @IsNotEmpty()
-  readonly accepted: boolean
-}
