--- backend/apps/client/src/contest/interface/publicizing-request.interface.ts
@@ -1,5 +1,0 @@
-export interface StoredPublicizingRequest {
-  contest: number
-  user: number
-  createTime: Date
-}
