+++ backend/apps/client/src/problem/workbook-problem.controller.ts
@@ -1,0 +1,119 @@
+import {
+  Controller,
+  Get,
+  InternalServerErrorException,
+  Logger,
+  NotFoundException,
+  Param,
+  ParseIntPipe,
+  Query,
+  UseGuards
+} from '@nestjs/common'
+import { AuthNotNeeded, GroupMemberGuard } from '@libs/auth'
+import { EntityNotExistException } from '@libs/exception'
+import { CursorValidationPipe } from '@libs/pipe'
+import { WorkbookProblemService } from './problem.service'
+
+@AuthNotNeeded()
+@Controller('workbook/:workbookId/problem')
+export class WorkbookProblemController {
+  private readonly logger = new Logger(WorkbookProblemController.name)
+
+  constructor(
+    private readonly workbookProblemService: WorkbookProblemService
+  ) {}
+
+  @Get()
+  async getWorkbookProblems(
+    @Param('workbookId', ParseIntPipe) workbookId: number,
+    @Query('cursor', CursorValidationPipe) cursor: number,
+    @Query('take', ParseIntPipe) take: number
+  ) {
+    try {
+      return await this.workbookProblemService.getWorkbookProblems(
+        workbookId,
+        cursor,
+        take
+      )
+    } catch (err) {
+      if (err instanceof EntityNotExistException) {
+        throw new NotFoundException(err.message)
+      }
+      this.logger.error(err.message, err.stack)
+      throw new InternalServerErrorException()
+    }
+  }
+
+  @Get(':problemId')
+  async getWorkbookProblem(
+    @Param('workbookId', ParseIntPipe) workbookId: number,
+    @Param('problemId', ParseIntPipe) problemId: number
+  ) {
+    try {
+      return await this.workbookProblemService.getWorkbookProblem(
+        workbookId,
+        problemId
+      )
+    } catch (err) {
+      if (err instanceof EntityNotExistException) {
+        throw new NotFoundException(err.message)
+      }
+      this.logger.error(err.message, err.stack)
+      throw new InternalServerErrorException()
+    }
+  }
+}
+
+@Controller('group/:groupId/workbook/:workbookId/problem')
+@UseGuards(GroupMemberGuard)
+export class GroupWorkbookProblemController {
+  private readonly logger = new Logger(GroupWorkbookProblemController.name)
+
+  constructor(
+    private readonly workbookProblemService: WorkbookProblemService
+  ) {}
+
+  @Get()
+  async getWorkbookProblems(
+    @Param('groupId', ParseIntPipe) groupId: number,
+    @Param('workbookId', ParseIntPipe) workbookId: number,
+    @Query('cursor', CursorValidationPipe) cursor: number,
+    @Query('take', ParseIntPipe) take: number
+  ) {
+    try {
+      return await this.workbookProblemService.getWorkbookProblems(
+        workbookId,
+        cursor,
+        take,
+        groupId
+      )
+    } catch (err) {
+      if (err instanceof EntityNotExistException) {
+        throw new NotFoundException(err.message)
+      }
+      this.logger.error(err.message, err.stack)
+      throw new InternalServerErrorException()
+    }
+  }
+
+  @Get(':problemId')
+  async getWorkbookProblem(
+    @Param('groupId', ParseIntPipe) groupId: number,
+    @Param('workbookId', ParseIntPipe) workbookId: number,
+    @Param('problemId', ParseIntPipe) problemId: number
+  ) {
+    try {
+      return await this.workbookProblemService.getWorkbookProblem(
+        workbookId,
+        problemId,
+        groupId
+      )
+    } catch (err) {
+      if (err instanceof EntityNotExistException) {
+        throw new NotFoundException(err.message)
+      }
+      this.logger.error(err.message, err.stack)
+      throw new InternalServerErrorException()
+    }
+  }
+}
